Project Not Loading Fix
This was a mistake on my part, I neglected to add the .csproj file to the commit earlier, meaning that the ContactDetails model and view were not loaded into the project. Since I refrenced the ContactDetails view in the Shell view, the project would compile and run but fail to get past the initial stages of loading.